# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sinapse Service Contracts
  version: 0.1.0
  description: Contratos versionados entre backend, serviço de IA e automações n8n.
servers:
  - url: http://localhost:3001
    description: Backend Node.js
  - url: http://localhost:8000
    description: Serviço de IA local
paths:
...
  /ingest/document:
    post:
      summary: Solicita ingestão assíncrona de documento
      operationId: ingestDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestDocumentRequest'
      responses:
        '200':
          description: Documento fragmentado e embeddings calculados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestDocumentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /ingest/entity:
    post:
      summary: Indexa uma entidade estruturada
      operationId: ingestEntity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestEntityRequest'
      responses:
        '200':
          description: Entidade processada
        '400':
          $ref: '#/components/responses/ValidationError'
  /embeddings:
    post:
      summary: Gera embedding para um texto
      operationId: generateEmbedding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmbeddingRequest'
      responses:
        '200':
          description: Vetor gerado
        '502':
          description: Ollama indisponível
  /rag/query:
    post:
      summary: Responde usando somente os trechos fornecidos
      operationId: queryRag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RagQueryRequest'
      responses:
        '200':
          description: Resposta fundamentada
        '502':
          description: Ollama indisponível
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sinapse_session
  schemas:
...
    IngestDocumentRequest:
      type: object
      required: [document_id, text_content]
      properties:
        document_id:
          type: string
        text_content:
          type: string
        project_id:
          type: string
          format: uuid
          nullable: true
        metadata:
          type: object
          additionalProperties: true
    IngestDocumentResponse:
      type: object
      required: [document_id, total_chunks, status, chunks]
      properties:
        document_id:
          type: string
        total_chunks:
          type: integer
          minimum: 0
        status:
          type: string
          enum: [chunked_and_indexed]
        chunks:
          type: array
          items:
            $ref: '#/components/schemas/ProcessedChunk'
    ProcessedChunk:
      type: object
      required: [chunk_index, text, vector_dimension]
      properties:
        chunk_index:
          type: integer
          minimum: 0
        text:
          type: string
        vector_dimension:
          type: integer
          minimum: 0
        project_id:
          type: string
          format: uuid
          nullable: true
    IngestEntityRequest:
      type: object
      required: [entity_type, data]
      properties:
        entity_type:
          type: string
          enum: [epic, feature, pbi, decision]
        data:
          type: object
          additionalProperties: true
    EmbeddingRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
        model:
          type: string
          nullable: true
    RagQueryRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
        project_id:
          type: string
          format: uuid
          nullable: true
        context_chunks:
          type: array
          items:
            type: string
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        correlation_id:
          type: string
          format: uuid
        details:
          type: object
          additionalProperties: true
  responses:
    ValidationError:
      description: Corpo inválido ou campo obrigatório ausente
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...